fix: making usePeople leaner as details and relationships aren't needed
